requirements.txt
streamlit
sentence-transformers
faiss-cpu
PyPDF2
python-dotenv
openai
